# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SyncSpace API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
  uptime: number;
}

export interface ErrorResponse {
  success: boolean;
  message: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface User {
  id: number;
  name: string;
  email: string;
  status: string;
  createdAt: string;
}

export interface AuthResponse {
  success: boolean;
  user: User;
  token: string;
}

export interface RegisterInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface Room {
  id: number;
  roomId: string;
  name: string;
  hostId: number;
  hostName: string;
  participantCount: number;
  isActive: boolean;
  createdAt: string;
  /** @nullable */
  endedAt: string | null;
}

export interface RoomInput {
  name: string;
}

export interface FileRecord {
  id: number;
  fileId: string;
  originalName: string;
  mimeType: string;
  size: number;
  uploadedBy: number;
  uploaderName: string;
  /** @nullable */
  roomId: string | null;
  createdAt: string;
}

export interface FileUploadInput {
  roomId?: string;
}

export interface AiSummarizeInput {
  messages: string[];
  roomId?: string;
}

export interface AiAskInput {
  question: string;
  context?: string;
}

export interface AiResponse {
  success: boolean;
  result: string;
  actionItems?: string[];
}

export type ListFilesParams = {
roomId?: string;
};