# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SyncSpace API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication
  - name: rooms
    description: Meeting rooms
  - name: files
    description: File sharing
  - name: ai
    description: AI assistant
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /rooms:
    post:
      operationId: createRoom
      tags: [rooms]
      summary: Create a new meeting room
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoomInput"
      responses:
        "201":
          description: Room created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /rooms/history:
    get:
      operationId: getRoomHistory
      tags: [rooms]
      summary: Get user's meeting history
      responses:
        "200":
          description: List of past rooms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Room"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /rooms/{roomId}:
    get:
      operationId: getRoom
      tags: [rooms]
      summary: Get room info
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Room info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
        "404":
          description: Room not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /files/upload:
    post:
      operationId: uploadFile
      tags: [files]
      summary: Upload a file
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/FileUploadInput"
      responses:
        "201":
          description: File uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileRecord"
        "400":
          description: Invalid file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /files:
    get:
      operationId: listFiles
      tags: [files]
      summary: List files (optionally filter by roomId)
      parameters:
        - name: roomId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of files
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FileRecord"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /files/{fileId}:
    get:
      operationId: getFile
      tags: [files]
      summary: Download a file
      parameters:
        - name: fileId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: File stream
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /ai/summarize:
    post:
      operationId: aiSummarize
      tags: [ai]
      summary: Summarize chat messages from a meeting
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiSummarizeInput"
      responses:
        "200":
          description: Summary result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /ai/ask:
    post:
      operationId: aiAsk
      tags: [ai]
      summary: Ask the AI assistant a question
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiAskInput"
      responses:
        "200":
          description: AI answer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
        uptime:
          type: number
      required:
        - status
        - uptime

    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success

    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - status
        - createdAt

    AuthResponse:
      type: object
      properties:
        success:
          type: boolean
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - success
        - user
        - token

    RegisterInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
      required:
        - name
        - email
        - password

    LoginInput:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password

    Room:
      type: object
      properties:
        id:
          type: integer
        roomId:
          type: string
        name:
          type: string
        hostId:
          type: integer
        hostName:
          type: string
        participantCount:
          type: integer
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        endedAt:
          type: ["string", "null"]
          format: date-time
      required:
        - id
        - roomId
        - name
        - hostId
        - hostName
        - participantCount
        - isActive
        - createdAt
        - endedAt

    RoomInput:
      type: object
      properties:
        name:
          type: string
      required:
        - name

    FileRecord:
      type: object
      properties:
        id:
          type: integer
        fileId:
          type: string
        originalName:
          type: string
        mimeType:
          type: string
        size:
          type: integer
        uploadedBy:
          type: integer
        uploaderName:
          type: string
        roomId:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - fileId
        - originalName
        - mimeType
        - size
        - uploadedBy
        - uploaderName
        - roomId
        - createdAt

    FileUploadInput:
      type: object
      properties:
        roomId:
          type: string

    AiSummarizeInput:
      type: object
      properties:
        messages:
          type: array
          items:
            type: string
        roomId:
          type: string
      required:
        - messages

    AiAskInput:
      type: object
      properties:
        question:
          type: string
        context:
          type: string
      required:
        - question

    AiResponse:
      type: object
      properties:
        success:
          type: boolean
        result:
          type: string
        actionItems:
          type: array
          items:
            type: string
      required:
        - success
        - result