# lib/api-zod/src/generated/types/aiAskInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SyncSpace API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AiAskInput {
  question: string;
  context?: string;
}