# lib/api-zod/src/generated/types/fileRecord.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SyncSpace API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FileRecord {
  id: number;
  fileId: string;
  originalName: string;
  mimeType: string;
  size: number;
  uploadedBy: number;
  uploaderName: string;
  /** @nullable */
  roomId: string | null;
  createdAt: Date;
}